ci: clear lint/security debt surfaced by first real CI run in 4.5 days
CI was quota-dead on the private repo since 2026-07-10, so lint/scan gates
couldn't run. Making the repo public re-enabled them and surfaced accumulated
debt across three workflows — all mechanical, no behavior change:

correlation-ci (ruff): unused imports (test_storm_release_gate via autofix),
a lambda→def (test_seam_kinds), an unused except var (azure), and a
section-local import noqa (cloud_log_parsers).

backend-ci (golangci-lint, repo-wide): ineffectual 'impact' init + unused
rcaIsImpactKind (both leftovers from the RCA-truthfulness merge), an
unnecessary uint64() (self_heal), and QF1008 embedded-field selectors
(path_graph_store Provenance, path_graph_isolation_test Spine — kept the one
genuinely-ambiguous sp.Spine.Spine range).

supply-chain (Trivy DS-0002): cloud-ingest poller runs root for host
bind-mount writes + /root/.aws creds — scoped .trivyignore exception with
justification; proper de-rooting tracked to Task #15 (Vault-scoped connectors).

go build/vet/test green; ruff + affected Go tests pass locally.

+++ NetOps_Observability/src/backend/self_heal.go
@@ -96,7 +96,7 @@ func diskUsedPct(path string) int {
 	if total == 0 {
 		return -1
 	}
-	return int(used * 100 / uint64(total))
+	return int(used * 100 / total)
 }
 
 // shouldHeal is the ONE decision rule (unit-tested): heal only when there is
